moving flags perser to another file
